implement getAccountNominatingData nominating util function
